Add possibility to select booking from ALLBookings page
